move select tests to Rest extension
